Python code to analyse the Epoch AI dataset
This program creates a visualisation of the number of citations per year for major AI publications related to games, language and vision respectively. It also assesses the proportion of the research that is conducted by academic groups compared to industry research. Finally, the program models the progression of the number of independent actors over time.